ADD: batch credential endpoint
