+++ src/main/java/es/in2/issuer/domain/model/BatchCredentialRequest.java
@@ -1,0 +1,12 @@
+package es.in2.issuer.domain.model;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.Builder;
+
+import java.util.List;
+
+@Builder
+public record BatchCredentialRequest(
+        @JsonProperty("credential_requests") List<CredentialRequest> credentialRequests
+        ) {
+}
