+++ src/main/java/es/in2/issuer/domain/model/CredentialConfiguration.java
@@ -9,8 +9,4 @@ public record CredentialConfiguration(@JsonProperty("format") String format,
                                       @JsonProperty("cryptographic_binding_methods_supported") List<String> cryptographicBindingMethodsSupported,
                                       @JsonProperty("credential_signing_alg_values_supported") List<String> credentialSigningAlgValuesSupported,
                                       @JsonProperty("credential_definition") CredentialDefinition credentialDefinition) {
-    @Builder
-    public record CredentialDefinition(
-            @JsonProperty("type") List<String> type
-    ) {}
 }
