+++ src/main/java/es/in2/issuer/domain/model/CredentialDefinition.java
@@ -1,0 +1,12 @@
+package es.in2.issuer.domain.model;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.Builder;
+
+import java.util.List;
+
+@Builder
+public record CredentialDefinition(
+        @JsonProperty("type") List<String> type
+) {
+}
