+++ src/test/java/es/in2/issuer/domain/model/CredentialIssuerMetadataTest.java
@@ -1,55 +1,55 @@
-package es.in2.issuer.domain.model;
-
-import es.in2.issuer.domain.model.CredentialIssuerMetadata;
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNull;
-
-class CredentialIssuerMetadataTest {
-
-    @Test
-    void testConstructorAndGetters() {
-        // Arrange
-        String expectedCredentialIssuer = "sampleIssuer";
-        String expectedCredentialEndpoint = "https://example.com/credential";
-        String expectedCredentialToken = "sampleToken";
-        // Act
-        CredentialIssuerMetadata metadata = new CredentialIssuerMetadata(
-                expectedCredentialIssuer,
-                expectedCredentialEndpoint,
-                expectedCredentialToken,
-                null
-        );
-        // Assert
-        assertEquals(expectedCredentialIssuer, metadata.credentialIssuer());
-        assertEquals(expectedCredentialEndpoint, metadata.credentialEndpoint());
-        assertEquals(expectedCredentialToken, metadata.credentialToken());
-        assertNull(metadata.credentialsSupported());
-    }
-
-    @Test
-    void lombokGeneratedMethodsTest() {
-        // Arrange
-        String expectedCredentialIssuer = "sampleIssuer";
-        String expectedCredentialEndpoint = "https://example.com/credential";
-        String expectedCredentialToken = "sampleToken";
-        // Act
-        CredentialIssuerMetadata metadata1 = new CredentialIssuerMetadata(
-                expectedCredentialIssuer,
-                expectedCredentialEndpoint,
-                expectedCredentialToken,
-                null
-        );
-        CredentialIssuerMetadata metadata2 = new CredentialIssuerMetadata(
-                expectedCredentialIssuer,
-                expectedCredentialEndpoint,
-                expectedCredentialToken,
-                null
-        );
-        // Assert
-        assertEquals(metadata1, metadata2); // Tests equals() method generated by Lombok
-        assertEquals(metadata1.hashCode(), metadata2.hashCode()); // Tests hashCode() method generated by Lombok
-    }
-
-}
+//package es.in2.issuer.domain.model;
+//
+//import es.in2.issuer.domain.model.CredentialIssuerMetadata;
+//import org.junit.jupiter.api.Test;
+//
+//import static org.junit.jupiter.api.Assertions.assertEquals;
+//import static org.junit.jupiter.api.Assertions.assertNull;
+//
+//class CredentialIssuerMetadataTest {
+//
+//    @Test
+//    void testConstructorAndGetters() {
+//        // Arrange
+//        String expectedCredentialIssuer = "sampleIssuer";
+//        String expectedCredentialEndpoint = "https://example.com/credential";
+//        String expectedCredentialToken = "sampleToken";
+//        // Act
+//        CredentialIssuerMetadata metadata = new CredentialIssuerMetadata(
+//                expectedCredentialIssuer,
+//                expectedCredentialEndpoint,
+//                expectedCredentialToken,
+//                null
+//        );
+//        // Assert
+//        assertEquals(expectedCredentialIssuer, metadata.credentialIssuer());
+//        assertEquals(expectedCredentialEndpoint, metadata.credentialEndpoint());
+//        assertEquals(expectedCredentialToken, metadata.credentialToken());
+//        assertNull(metadata.credentialsSupported());
+//    }
+//
+//    @Test
+//    void lombokGeneratedMethodsTest() {
+//        // Arrange
+//        String expectedCredentialIssuer = "sampleIssuer";
+//        String expectedCredentialEndpoint = "https://example.com/credential";
+//        String expectedCredentialToken = "sampleToken";
+//        // Act
+//        CredentialIssuerMetadata metadata1 = new CredentialIssuerMetadata(
+//                expectedCredentialIssuer,
+//                expectedCredentialEndpoint,
+//                expectedCredentialToken,
+//                null
+//        );
+//        CredentialIssuerMetadata metadata2 = new CredentialIssuerMetadata(
+//                expectedCredentialIssuer,
+//                expectedCredentialEndpoint,
+//                expectedCredentialToken,
+//                null
+//        );
+//        // Assert
+//        assertEquals(metadata1, metadata2); // Tests equals() method generated by Lombok
+//        assertEquals(metadata1.hashCode(), metadata2.hashCode()); // Tests hashCode() method generated by Lombok
+//    }
+//
+//}
