+++ src/test/java/es/in2/issuer/domain/model/CredentialRequestTest.java
@@ -1,34 +1,34 @@
-package es.in2.issuer.domain.model;
-
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
-class CredentialRequestTest {
-
-    @Test
-    void testConstructorAndGetters() {
-        // Arrange
-        String expectedFormat = "sampleFormat";
-        Proof expectedProof = new Proof("jwt_vc_json", "sampleJwt");
-        // Act
-        CredentialRequest credentialRequest = new CredentialRequest(expectedFormat, expectedProof);
-        // Assert
-        assertEquals(expectedFormat, credentialRequest.format());
-        assertEquals(expectedProof, credentialRequest.proof());
-    }
-
-    @Test
-    void lombokGeneratedMethodsTest() {
-        // Arrange
-        String expectedFormat = "sampleFormat";
-        Proof expectedProof = new Proof("jwt_vc_json", "sampleJwt");
-        // Act
-        CredentialRequest dto1 = new CredentialRequest(expectedFormat, expectedProof);
-        CredentialRequest dto2 = new CredentialRequest(expectedFormat, expectedProof);
-        // Assert
-        assertEquals(dto1, dto2); // Tests equals() method generated by Lombok
-        assertEquals(dto1.hashCode(), dto2.hashCode()); // Tests hashCode() method generated by Lombok
-    }
-
-}
+//package es.in2.issuer.domain.model;
+//
+//import org.junit.jupiter.api.Test;
+//
+//import static org.junit.jupiter.api.Assertions.assertEquals;
+//
+//class CredentialRequestTest {
+//
+//    @Test
+//    void testConstructorAndGetters() {
+//        // Arrange
+//        String expectedFormat = "sampleFormat";
+//        Proof expectedProof = new Proof("jwt_vc_json", "sampleJwt");
+//        // Act
+//        CredentialRequest credentialRequest = new CredentialRequest(expectedFormat, expectedProof);
+//        // Assert
+//        assertEquals(expectedFormat, credentialRequest.format());
+//        assertEquals(expectedProof, credentialRequest.proof());
+//    }
+//
+//    @Test
+//    void lombokGeneratedMethodsTest() {
+//        // Arrange
+//        String expectedFormat = "sampleFormat";
+//        Proof expectedProof = new Proof("jwt_vc_json", "sampleJwt");
+//        // Act
+//        CredentialRequest dto1 = new CredentialRequest(expectedFormat, expectedProof);
+//        CredentialRequest dto2 = new CredentialRequest(expectedFormat, expectedProof);
+//        // Assert
+//        assertEquals(dto1, dto2); // Tests equals() method generated by Lombok
+//        assertEquals(dto1.hashCode(), dto2.hashCode()); // Tests hashCode() method generated by Lombok
+//    }
+//
+//}
